app/agents/meta_agents/integration_agent.py
from typing import Any, Dict, List, Tuple
import asyncio
from app.agents.meta_agents.meta_agents import MetaAgent
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.ontology_reasoner import OntologyReasoner
from app.models.knowledge.knowledge_base import KnowledgeBase
from app.tools.llm_manager import LLMManager
from app.models.knowledge.ontology.ontology import Ontology
from pydantic import BaseModel

class IntegrationAgent(MetaAgent):
    def __init__(self, agent_id: str, name: str, api_key: str, llm_service: Any, agent_communication_service: Any):
        super().__init__(
            agent_id=agent_id,
            name=name,
            api_key=api_key,
            llm_service=llm_service,
            agent_communication_service=agent_communication_service
        )
        self.agent_type = "integration"
        self.knowledge_base = KnowledgeBase()
        self.reasoning_engine = ReasoningEngine(self.knowledge_base, api_key)
        self.llm_manager = LLMManager()
        self.ontology = Ontology()  # Initialize with your integration ontology
        self.ontology_reasoner = OntologyReasoner(self.llm_manager, self.ontology)
        self.communication_protocols = {}

        self._init_beliefs()
        self._init_desires()
        self._init_goals()
        self._init_plans()

    def _init_beliefs(self):
        self.add_belief("Effective integration is crucial for MAS performance")
        self.add_belief("Seamless communication between agents enhances system efficiency")

    def _init_desires(self):
        self.add_desire("Ensure smooth integration of all MAS components", priority=10)
        self.add_desire("Optimize inter-agent communication", priority=9)

    def _init_goals(self):
        self.add_goal("Integrate agent subsystems", priority=9)
        self.add_goal("Implement robust inter-agent communication", priority=8)
        self.add_goal("Resolve integration conflicts", priority=8)

    def _init_plans(self):
        self.create_plan(
            self.goals[0].id,
            [
                "Analyze agent subsystems",
                "Identify integration points",
                "Develop integration strategies",
                "Implement integration solutions",
                "Test integrated components"
            ]
        )

    async def reason(self):
        print("Starting reasoning process for integration")
        
        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        updated_beliefs = await self.reasoning_engine.reason({"beliefs": current_beliefs})
        
        for belief in updated_beliefs.get("beliefs", []):
            self.add_belief(belief["content"])

        new_knowledge = await self.ontology_reasoner.infer_new_knowledge()
        
        for concept in new_knowledge.get("new_concepts", []):
            self.add_belief(f"New integration concept: {concept['name']} - {concept['description']}")
        
        for relationship in new_knowledge.get("new_relationships", []):
            self.add_belief(f"New integration relationship: {relationship['name']} between {relationship['domain']} and {relationship['range']}")

        integration_query = "What are the best practices for integrating MAS components?"
        integration_practices = await self.ontology_reasoner.answer_query(integration_query)
        self.add_belief(f"Best integration practices: {integration_practices}")

        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        new_desires = await self.reasoning_engine.generate_desires(current_beliefs)
        
        for desire in new_desires:
            self.add_desire(desire["description"], desire["priority"])

        print("Reasoning process for integration completed")

    async def plan(self):
        print("Starting planning process for integration")
        for goal in self.goals:
            if goal.description == "Implement robust inter-agent communication":
                self.create_plan(
                    goal.id,
                    [
                        "Analyze communication requirements",
                        "Design communication protocols",
                        "Implement communication interfaces",
                        "Test inter-agent communication",
                        "Optimize communication performance"
                    ]
                )
        
        current_state = self.get_current_state()
        optimized_plan = await self.reasoning_engine.reason_and_plan(self.goals[0].description, current_state)
        
        if optimized_plan.get("plan"):
            self.update_plan(self.goals[0].id, optimized_plan["plan"].steps)
        
        print("Planning process for integration completed")

    async def execute(self):
        print("Starting execution process for integration")
        for plan in self.plans:
            for task in plan.steps:
                if task.status == "pending":
                    print(f"Executing task: {task.description}")
                    try:
                        execution_result = await self.reasoning_engine.simulate_action(task.description, self.get_current_state())
                        self.update_task_status(task.id, "completed")
                        
                        for key, value in execution_result.items():
                            self.add_belief(f"Integration task result - {key}: {value}")
                    except Exception as e:
                        print(f"Error executing integration task {task.description}: {str(e)}")
                        self.update_task_status(task.id, "failed")
        print("Execution process for integration completed")

    def get_current_state(self) -> Dict[str, Any]:
        return {
            "beliefs": [belief.to_dict() for belief in self.beliefs],
            "desires": [desire.to_dict() for desire in self.desires],
            "goals": [goal.to_dict() for goal in self.goals],
            "plans": [plan.to_dict() for plan in self.plans],
        }

    async def run(self):
        while True:
            await self.reason()
            await self.plan()
            await self.execute()
            await asyncio.sleep(1)  # Adjust the sleep time as needed

    async def integrate_agent_subsystems(self, agent_subsystems: Dict[str, Any]) -> Dict[str, Any]:
        print("Integrating agent subsystems...")
        integration_tasks = [
            self._integrate_subsystem(subsystem, details)
            for subsystem, details in agent_subsystems.items()
        ]
        
        results = await asyncio.gather(*integration_tasks)
        return dict(results)

    async def _integrate_subsystem(self, subsystem: str, details: Any) -> Tuple[str, Any]:
        print(f"Integrating subsystem: {subsystem}")
        # Implement integration logic for each subsystem
        # This is a placeholder implementation
        integration_result = await self.reasoning_engine.simulate_action(f"integrate_{subsystem}", details)
        return (subsystem, integration_result)

    async def implement_inter_agent_communication(self, agent_interactions: Dict[str, Any]) -> Dict[str, Any]:
        print("Implementing inter-agent communication...")
        for interaction, details in agent_interactions.items():
            protocol = await self._implement_protocol(details)
            self.communication_protocols[interaction] = protocol
        return self.communication_protocols

    async def _implement_protocol(self, details: Dict[str, Any]):
        print(f"Implementing communication protocol: {details.get('name', 'Unnamed')}")
        # Implement secure communication protocol
        # This is a placeholder implementation
        protocol_implementation = await self.reasoning_engine.simulate_action("implement_protocol", details)
        return protocol_implementation

    async def resolve_conflicts(self, conflicting_agents: List[str], conflict_details: Dict[str, Any]) -> Dict[str, Any]:
        print("Resolving conflicts between agents...")
        resolution = {}
        for agent in conflicting_agents:
            resolution[agent] = await self._negotiate_resolution(agent, conflict_details)
        return resolution

    async def _negotiate_resolution(self, agent: str, conflict_details: Dict[str, Any]) -> Dict[str, Any]:
        print(f"Negotiating resolution for agent: {agent}")
        # Implement conflict resolution logic
        # This is a placeholder implementation
        resolution = await self.reasoning_engine.simulate_action("negotiate_resolution", {
            "agent": agent,
            "conflict": conflict_details
        })
        return resolution

    async def optimize_integration(self, current_integration: Dict[str, Any]) -> Dict[str, Any]:
        print("Optimizing integration...")
        optimization_query = f"How can we optimize this integration: {current_integration}?"
        optimization_suggestions = await self.ontology_reasoner.answer_query(optimization_query)
        
        optimized_integration = current_integration.copy()
        for suggestion in optimization_suggestions:
            optimized_integration = await self._apply_optimization(optimized_integration, suggestion)
        
        return optimized_integration

    async def _apply_optimization(self, integration: Dict[str, Any], optimization: str) -> Dict[str, Any]:
        print(f"Applying optimization: {optimization}")
        # Implement optimization logic
        # This is a placeholder implementation
        optimized = await self.reasoning_engine.simulate_action("apply_optimization", {
            "integration": integration,
            "optimization": optimization
        })
        return optimized
