tests/unit/test_integration_agent.py
import unittest
from unittest.mock import patch, MagicMock
from app.agents.meta_agents.integration_agent import IntegrationAgent

class TestIntegrationAgent(unittest.TestCase):

    def setUp(self):
        self.agent = IntegrationAgent()

    @patch('app.agents.meta_agents.integration_agent.IntegrationAgent._integrate_subsystem')
    def test_integrate_agent_subsystems(self, mock_integrate_subsystem):
        mock_integrate_subsystem.return_value = ("subsystem1", "details1")
        agent_subsystems = {"subsystem1": "details1"}
        result = self.agent.integrate_agent_subsystems(agent_subsystems)
        self.assertEqual(result, {"subsystem1": "details1"})

    def test_implement_inter_agent_communication(self):
        agent_interactions = {"interaction1": {"details": "details1"}}
        result = self.agent.implement_inter_agent_communication(agent_interactions)
        self.assertIn("interaction1", result)

    def test_resolve_conflicts(self):
        conflicting_agents = ["agent1", "agent2"]
        conflict_details = {"agent1": "details1", "agent2": "details2"}
        result = self.agent.resolve_conflicts(conflicting_agents, conflict_details)
        self.assertIn("agent1", result)
        self.assertIn("agent2", result)

if __name__ == '__main__':
    unittest.main()